+++ pkg/util/k8sutil/helm/chart_manager_test.go
@@ -92,3 +92,13 @@ func Test_Manager(t *testing.T) {
 
 	require.NoError(t, mgr.Reload(context.Background()))
 }
+
+func Test_Manager_Tag(t *testing.T) {
+	mgr, err := NewChartManager(context.Background(), nil, "https://arangodb-platform-dev-chart-registry.s3.amazonaws.com/index.yaml")
+	require.NoError(t, err)
+
+	repo, ok := mgr.Get("platform_test_example")
+	require.True(t, ok)
+
+	repo.GetByTag("dev")
+}